refactor(github): split fleet sync into focused modules
Extract fleet sync into dedicated modules, share Git helpers with tidy, and add parser standardization coverage for validation CLI usage.

Closes #1406

Implemented-By: Codex
