+++ hephaestus/github/fleet_sync/__init__.py
@@ -1,0 +1,71 @@
+"""Fleet-wide PR synchronization CLI and helpers."""
+
+from __future__ import annotations
+
+from hephaestus.github.fleet_sync.cli import _build_parser, main
+from hephaestus.github.fleet_sync.config import (
+    DEFAULT_FLEET_CONFIG_FILENAME,
+    _find_default_config,
+    _load_fleet_config,
+    resolve_fleet_config,
+)
+from hephaestus.github.fleet_sync.conflict_resolver import (
+    _run_conflict_agent,
+    resolve_conflict_with_agent,
+)
+from hephaestus.github.fleet_sync.git_ops import (
+    _git,
+    add_pr_worktree,
+    ensure_repo_clone,
+    rebase_and_resign,
+    remove_worktree,
+)
+from hephaestus.github.fleet_sync.gpg import get_resign_email, get_resign_exec
+from hephaestus.github.fleet_sync.models import (
+    ASCII_SYMBOLS,
+    UNICODE_SYMBOLS,
+    PRInfo,
+    PRStatus,
+    Symbols,
+)
+from hephaestus.github.fleet_sync.pr_api import (
+    _ci_state,
+    _fetch_pr_ci_state,
+    _gh,
+    list_prs,
+    merge_pr,
+)
+from hephaestus.github.fleet_sync.sync_coordinator import logger, process_repo
+from hephaestus.utils.helpers import METADATA_TIMEOUT, NETWORK_TIMEOUT
+
+__all__ = [
+    "ASCII_SYMBOLS",
+    "DEFAULT_FLEET_CONFIG_FILENAME",
+    "METADATA_TIMEOUT",
+    "NETWORK_TIMEOUT",
+    "UNICODE_SYMBOLS",
+    "PRInfo",
+    "PRStatus",
+    "Symbols",
+    "_build_parser",
+    "_ci_state",
+    "_fetch_pr_ci_state",
+    "_find_default_config",
+    "_gh",
+    "_git",
+    "_load_fleet_config",
+    "_run_conflict_agent",
+    "add_pr_worktree",
+    "ensure_repo_clone",
+    "get_resign_email",
+    "get_resign_exec",
+    "list_prs",
+    "logger",
+    "main",
+    "merge_pr",
+    "process_repo",
+    "rebase_and_resign",
+    "remove_worktree",
+    "resolve_conflict_with_agent",
+    "resolve_fleet_config",
+]
