+++ tests/unit/github/test_fleet_sync.py
@@ -2,17 +2,26 @@
 
 from __future__ import annotations
 
-import importlib
 import json
 import subprocess
 from pathlib import Path
+from typing import Any, cast
 from unittest.mock import MagicMock, patch
 
 import pytest
 
-from hephaestus.github.fleet_sync import PRInfo, PRStatus, _ci_state, get_resign_email
-
-fleet_sync_module = importlib.import_module("hephaestus.github.fleet_sync")
+from hephaestus.github.fleet_sync import (
+    cli as fleet_cli,
+    conflict_resolver as fleet_conflicts,
+    git_ops as fleet_git_ops,
+    gpg as fleet_gpg,
+    models as fleet_models,
+    pr_api as fleet_pr_api,
+    sync_coordinator as fleet_coordinator,
+)
+from hephaestus.github.fleet_sync.models import PRInfo, PRStatus
+from hephaestus.github.fleet_sync.pr_api import _ci_state
+from hephaestus.utils.helpers import METADATA_TIMEOUT, NETWORK_TIMEOUT
 
 
 class TestCiState:
@@ -183,11 +192,9 @@ class TestGetResignEmail:
 
     def test_env_var_takes_precedence(self, monkeypatch) -> None:
         """FLEET_GIT_EMAIL is used when set."""
-        from hephaestus.github.fleet_sync import get_resign_email
-
         monkeypatch.setenv("FLEET_SKIP_EMAIL_KEY_CHECK", "1")
         monkeypatch.setenv("FLEET_GIT_EMAIL", "alice@example.com")
-        assert get_resign_email() == "alice@example.com"
+        assert fleet_gpg.get_resign_email() == "alice@example.com"
 
     def test_empty_env_var_falls_through_to_git_config(self, monkeypatch) -> None:
         """An empty FLEET_GIT_EMAIL falls back to git config."""
@@ -206,7 +213,7 @@ class TestGetResignEmail:
         # Target the attribute by dotted path so strict mypy (implicit_reexport=False)
         # doesn't complain about fleet_sync not re-exporting `subprocess`.
         monkeypatch.setattr(
-            "hephaestus.github.fleet_sync.subprocess.run",
+            "hephaestus.github.fleet_sync.gpg.subprocess.run",
             lambda *a, **k: _Result(),
         )
         assert fleet_sync.get_resign_email() == "bob@example.com"
@@ -224,7 +231,7 @@ class TestGetResignEmail:
             stdout = ""
 
         monkeypatch.setattr(
-            "hephaestus.github.fleet_sync.subprocess.run",
+            "hephaestus.github.fleet_sync.gpg.subprocess.run",
             lambda *a, **k: _EmptyResult(),
         )
         with pytest.raises(RuntimeError, match="no resign email configured"):
@@ -267,7 +274,7 @@ class TestResignEmailKeyGuard:
                 result.stdout = ""
             return result
 
-        monkeypatch.setattr("hephaestus.github.fleet_sync.subprocess.run", fake_run)
+        monkeypatch.setattr("hephaestus.github.fleet_sync.gpg.subprocess.run", fake_run)
 
     def test_email_on_key_is_accepted(self, monkeypatch) -> None:
         """Resolution succeeds when the email is a UID on the signing key."""
@@ -321,7 +328,7 @@ class TestResignEmailKeyGuard:
             result.stdout = "ABC123\n"
             return result
 
-        monkeypatch.setattr("hephaestus.github.fleet_sync.subprocess.run", fake_run)
+        monkeypatch.setattr("hephaestus.github.fleet_sync.gpg.subprocess.run", fake_run)
         assert fleet_sync.get_resign_email() == "anything@example.com"
 
 
@@ -330,7 +337,6 @@ class TestMain:
 
     def test_main_success_json(self, monkeypatch, capsys) -> None:
         """main() with --json emits ok envelope when no failures occur."""
-        from hephaestus.github import fleet_sync
 
         def fake_process(repo, org, args, clone_dir, *, symbols=None):
             return {
@@ -341,7 +347,7 @@ class TestMain:
                 "failed": 0,
             }
 
-        monkeypatch.setattr(fleet_sync, "process_repo", fake_process)
+        monkeypatch.setattr(fleet_cli, "process_repo", fake_process)
         monkeypatch.setattr(
             "sys.argv",
             [
@@ -356,7 +362,7 @@ class TestMain:
                 "claude",
             ],
         )
-        assert fleet_sync.main() == 0
+        assert fleet_cli.main() == 0
         payload = json.loads(capsys.readouterr().out)
         assert payload["status"] == "ok"
         assert payload["repos"] == 1
@@ -364,7 +370,6 @@ class TestMain:
 
     def test_main_failure_json(self, monkeypatch, capsys) -> None:
         """main() with failures returns 1 and JSON envelope shows error."""
-        from hephaestus.github import fleet_sync
 
         def fake_process(repo, org, args, clone_dir, *, symbols=None):
             return {
@@ -375,7 +380,7 @@ class TestMain:
                 "failed": 1,
             }
 
-        monkeypatch.setattr(fleet_sync, "process_repo", fake_process)
+        monkeypatch.setattr(fleet_cli, "process_repo", fake_process)
         monkeypatch.setattr(
             "sys.argv",
             [
@@ -390,15 +395,13 @@ class TestMain:
                 "claude",
             ],
         )
-        assert fleet_sync.main() == 1
+        assert fleet_cli.main() == 1
         payload = json.loads(capsys.readouterr().out)
         assert payload["status"] == "error"
         assert payload["totals"]["failed"] == 1
 
     def test_main_success_text(self, monkeypatch) -> None:
         """main() without --json still runs through every repo."""
-        from hephaestus.github import fleet_sync
-
         calls = []
 
         def fake_process(repo, org, args, clone_dir, *, symbols=None):
@@ -411,12 +414,12 @@ class TestMain:
                 "failed": 0,
             }
 
-        monkeypatch.setattr(fleet_sync, "process_repo", fake_process)
+        monkeypatch.setattr(fleet_cli, "process_repo", fake_process)
         monkeypatch.setattr(
             "sys.argv",
             ["fleet-sync", "--org", "owner", "--repos", "a", "b", "--dry-run", "--agent", "claude"],
         )
-        assert fleet_sync.main() == 0
+        assert fleet_cli.main() == 0
         assert calls == ["a", "b"]
 
 
@@ -441,9 +444,9 @@ class TestTimeoutHandling:
             result.stdout = "alice@example.com\n"
             return result
 
-        monkeypatch.setattr("hephaestus.github.fleet_sync.subprocess.run", failing_run)
+        monkeypatch.setattr("hephaestus.github.fleet_sync.gpg.subprocess.run", failing_run)
         # Should get email from second attempt (after timeout)
-        assert get_resign_email() == "alice@example.com"
+        assert fleet_gpg.get_resign_email() == "alice@example.com"
 
     def test_get_resign_email_uses_metadata_timeout(self, monkeypatch: pytest.MonkeyPatch) -> None:
         """get_resign_email uses METADATA_TIMEOUT."""
@@ -453,23 +456,23 @@ class TestTimeoutHandling:
 
         with patch("subprocess.run") as mock_run:
             mock_run.return_value = MagicMock(returncode=0, stdout="test@example.com\n")
-            get_resign_email()
+            fleet_gpg.get_resign_email()
             assert mock_run.called
             call_kwargs = mock_run.call_args[1]
             assert "timeout" in call_kwargs
-            assert call_kwargs["timeout"] == fleet_sync_module.METADATA_TIMEOUT
+            assert call_kwargs["timeout"] == METADATA_TIMEOUT
 
     def test_gh_delegates_to_gh_call_and_scopes_repo(self) -> None:
         """_gh routes through gh_call with repo scoping and NETWORK_TIMEOUT."""
         completed = subprocess.CompletedProcess(["gh"], 0, stdout="[]", stderr="")
-        with patch.object(fleet_sync_module, "gh_call", return_value=completed) as mock_gh_call:
-            result = fleet_sync_module._gh(["pr", "list"], repo="TestRepo", org="TestOrg")
+        with patch.object(fleet_pr_api, "gh_call", return_value=completed) as mock_gh_call:
+            result = fleet_pr_api._gh(["pr", "list"], repo="TestRepo", org="TestOrg")
 
         assert result is completed
         mock_gh_call.assert_called_once_with(
             ["--repo", "TestOrg/TestRepo", "pr", "list"],
             check=True,
-            timeout=fleet_sync_module.NETWORK_TIMEOUT,
+            timeout=NETWORK_TIMEOUT,
         )
 
     def test_git_uses_network_timeout(self) -> None:
@@ -477,21 +480,21 @@ class TestTimeoutHandling:
         with patch("subprocess.run") as mock_run:
             mock_run.return_value = MagicMock(stdout="")
             work_dir = Path("/tmp/test")
-            fleet_sync_module._git(["clone", "url", "."], cwd=work_dir)
+            fleet_git_ops._git(["clone", "url", "."], cwd=work_dir)
             assert mock_run.called
             call_kwargs = mock_run.call_args[1]
             assert "timeout" in call_kwargs
-            assert call_kwargs["timeout"] == fleet_sync_module.NETWORK_TIMEOUT
+            assert call_kwargs["timeout"] == NETWORK_TIMEOUT
 
     def test_conflict_agent_uses_env_configured_rebase_timeout(
         self, monkeypatch: pytest.MonkeyPatch
     ) -> None:
         """Direct fleet conflict agents use the centralized rebase timeout."""
         monkeypatch.setenv("HEPH_AGENT_REBASE_TIMEOUT", "1234")
-        with patch("hephaestus.github.fleet_sync.run_agent_text") as run_agent:
+        with patch("hephaestus.github.fleet_sync.conflict_resolver.run_agent_text") as run_agent:
             run_agent.return_value = MagicMock(stdout="resolved")
 
-            assert fleet_sync_module._run_conflict_agent("codex", "prompt", Path("/repo"), 7)
+            assert fleet_conflicts._run_conflict_agent("codex", "prompt", Path("/repo"), 7)
 
         assert run_agent.call_args.kwargs["timeout"] == 1234
 
@@ -540,7 +543,7 @@ def capture_fleet_sync_logs():
                 messages.append(record.getMessage())
 
         # Get the underlying logger (ContextLogger wraps a LoggerAdapter)
-        logger_adapter = fleet_sync_module.logger
+        logger_adapter = fleet_coordinator.logger
         underlying_logger = logger_adapter.logger
         handler = _TestHandler()
         underlying_logger.addHandler(handler)
@@ -563,8 +566,8 @@ class TestCloneReuseAndWorktrees:
             calls.append(args)
             return MagicMock(returncode=0, stdout="", stderr="")
 
-        with patch.object(fleet_sync_module, "_git", side_effect=fake_git):
-            path = fleet_sync_module.ensure_repo_clone("RepoA", "HomericIntelligence", tmp_path)
+        with patch.object(fleet_git_ops, "_git", side_effect=fake_git):
+            path = fleet_git_ops.ensure_repo_clone("RepoA", "HomericIntelligence", tmp_path)
 
         assert path == tmp_path / "RepoA"
         assert calls[0][0] == "clone"
@@ -579,8 +582,8 @@ class TestCloneReuseAndWorktrees:
             calls.append(args)
             return MagicMock(returncode=0, stdout="", stderr="")
 
-        with patch.object(fleet_sync_module, "_git", side_effect=fake_git):
-            path = fleet_sync_module.ensure_repo_clone("RepoA", "HomericIntelligence", tmp_path)
+        with patch.object(fleet_git_ops, "_git", side_effect=fake_git):
+            path = fleet_git_ops.ensure_repo_clone("RepoA", "HomericIntelligence", tmp_path)
 
         assert path == tmp_path / "RepoA"
         # Reuse path fetches, never clones.
@@ -596,8 +599,8 @@ class TestCloneReuseAndWorktrees:
             calls.append(args)
             return MagicMock(returncode=0, stdout="", stderr="")
 
-        with patch.object(fleet_sync_module, "_git", side_effect=fake_git):
-            fleet_sync_module.add_pr_worktree(repo_clone, work, "feat", "main")
+        with patch.object(fleet_git_ops, "_git", side_effect=fake_git):
+            fleet_git_ops.add_pr_worktree(repo_clone, work, "feat", "main")
 
         assert ["fetch", "origin", "feat"] in calls
         assert ["fetch", "origin", "main"] in calls
@@ -619,10 +622,10 @@ class TestCloneReuseAndWorktrees:
             return MagicMock(returncode=0, stdout="", stderr="")
 
         with (
-            patch.object(fleet_sync_module, "_git", side_effect=fake_git),
-            patch.object(fleet_sync_module, "get_resign_exec", return_value="true"),
+            patch.object(fleet_git_ops, "_git", side_effect=fake_git),
+            patch.object(fleet_git_ops, "get_resign_exec", return_value="true"),
         ):
-            ok = fleet_sync_module.rebase_and_resign(pr, repo_clone)
+            ok = fleet_git_ops.rebase_and_resign(pr, repo_clone)
 
         assert ok is True
         # Never clones; uses a worktree and cleans it up.
@@ -642,11 +645,11 @@ class TestCloneReuseAndWorktrees:
         args = MagicMock(dry_run=False, skip_conflict_resolution=False, agent="claude")
 
         with (
-            patch.object(fleet_sync_module, "list_prs", return_value=prs),
-            patch.object(fleet_sync_module, "ensure_repo_clone", side_effect=fake_ensure),
-            patch.object(fleet_sync_module, "rebase_and_resign", return_value=True),
+            patch.object(fleet_coordinator, "list_prs", return_value=prs),
+            patch.object(fleet_coordinator, "ensure_repo_clone", side_effect=fake_ensure),
+            patch.object(fleet_coordinator, "rebase_and_resign", return_value=True),
         ):
-            counts = fleet_sync_module.process_repo("RepoA", "HomericIntelligence", args, tmp_path)
+            counts = fleet_coordinator.process_repo("RepoA", "HomericIntelligence", args, tmp_path)
 
         assert clone_count[0] == 1
         assert counts["rebased"] == 3
@@ -663,11 +666,11 @@ class TestCloneReuseAndWorktrees:
         args = MagicMock(dry_run=False, skip_conflict_resolution=False, agent="claude")
 
         with (
-            patch.object(fleet_sync_module, "list_prs", return_value=prs),
-            patch.object(fleet_sync_module, "ensure_repo_clone", side_effect=fake_ensure),
-            patch.object(fleet_sync_module, "merge_pr", return_value=True),
+            patch.object(fleet_coordinator, "list_prs", return_value=prs),
+            patch.object(fleet_coordinator, "ensure_repo_clone", side_effect=fake_ensure),
+            patch.object(fleet_coordinator, "merge_pr", return_value=True),
         ):
-            counts = fleet_sync_module.process_repo("RepoA", "HomericIntelligence", args, tmp_path)
+            counts = fleet_coordinator.process_repo("RepoA", "HomericIntelligence", args, tmp_path)
 
         assert clone_count[0] == 0
         assert counts["merged"] == 1
@@ -680,8 +683,8 @@ class TestMergePr:
         """Circuit-breaker/runtime failures return False like gh CLI failures."""
         pr = _pr(42, PRStatus.READY)
 
-        with patch.object(fleet_sync_module, "_gh", side_effect=RuntimeError("breaker open")):
-            assert fleet_sync_module.merge_pr(pr, "HomericIntelligence") is False
+        with patch.object(fleet_pr_api, "_gh", side_effect=RuntimeError("breaker open")):
+            assert fleet_pr_api.merge_pr(pr, "HomericIntelligence") is False
 
 
 class TestListPrs:
@@ -717,8 +720,8 @@ class TestListPrs:
             # pr view (per-PR CI fetch)
             return MagicMock(stdout=json.dumps({"statusCheckRollup": []}))
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
-        prs = fleet_sync_module.list_prs("ProjectHephaestus", "HomericIntelligence")
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
+        prs = fleet_pr_api.list_prs("ProjectHephaestus", "HomericIntelligence")
         json_idx = captured["list_args"].index("--json")
         json_fields = captured["list_args"][json_idx + 1]
         assert "statusCheckRollup" not in json_fields
@@ -753,8 +756,8 @@ class TestListPrs:
                 )
             )
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
-        prs = fleet_sync_module.list_prs("ProjectHephaestus", "HomericIntelligence")
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
+        prs = fleet_pr_api.list_prs("ProjectHephaestus", "HomericIntelligence")
         assert view_calls and view_calls[0][:3] == ["pr", "view", "7"]
         assert prs[0].ci_state == "SUCCESS"
         assert prs[0].status == PRStatus.READY
@@ -781,8 +784,8 @@ class TestListPrs:
                 )
             raise RuntimeError("504 on pr view")
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
-        prs = fleet_sync_module.list_prs("ProjectHephaestus", "HomericIntelligence")
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
+        prs = fleet_pr_api.list_prs("ProjectHephaestus", "HomericIntelligence")
         assert prs[0].ci_state == "UNKNOWN"
 
     def test_list_failure_raises_not_swallowed(self, monkeypatch: pytest.MonkeyPatch) -> None:
@@ -791,9 +794,9 @@ class TestListPrs:
         def fake_gh(args, repo=None, **kwargs):
             raise subprocess.CalledProcessError(1, ["gh"], stderr="HTTP 504")
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
         with pytest.raises(RuntimeError, match="could not list PRs"):
-            fleet_sync_module.list_prs("ProjectHephaestus", "HomericIntelligence")
+            fleet_pr_api.list_prs("ProjectHephaestus", "HomericIntelligence")
 
     def test_empty_list_returns_empty(self, monkeypatch: pytest.MonkeyPatch) -> None:
         """An actually-empty repo returns [] (distinct from a list failure)."""
@@ -801,8 +804,8 @@ class TestListPrs:
         def fake_gh(args, repo=None, **kwargs):
             return MagicMock(stdout="[]")
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
-        assert fleet_sync_module.list_prs("ProjectHephaestus", "HomericIntelligence") == []
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
+        assert fleet_pr_api.list_prs("ProjectHephaestus", "HomericIntelligence") == []
 
 
 class TestPrClassification:
@@ -843,8 +846,8 @@ class TestPrClassification:
                 rollup = [{"conclusion": "SUCCESS", "state": "SUCCESS"}]
             return MagicMock(stdout=json.dumps({"statusCheckRollup": rollup}))
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
-        return fleet_sync_module.list_prs("ProjectHephaestus", "HomericIntelligence")[0].status
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
+        return fleet_pr_api.list_prs("ProjectHephaestus", "HomericIntelligence")[0].status
 
     def test_blocked_mergeable_failing_is_outdated(self, monkeypatch) -> None:
         """BLOCKED+MERGEABLE with red CI = stale failure → rebase (OUTDATED)."""
@@ -900,9 +903,9 @@ class TestListPrsAuthorScope:
             captured_argv.append(args)
             return MagicMock(stdout="[]")
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
 
-        fleet_sync_module.list_prs("RepoA", "HomericIntelligence")
+        fleet_pr_api.list_prs("RepoA", "HomericIntelligence")
 
         assert captured_argv, "_gh was never called"
         pr_list_argv = captured_argv[0]
@@ -919,7 +922,7 @@ class TestListPrsAuthorScope:
         separate client-side filtering can drift out of sync.
         """
         monkeypatch.setattr(
-            fleet_sync_module,
+            fleet_pr_api,
             "_fetch_pr_ci_state",
             lambda repo, number, org=None: "SUCCESS",
         )
@@ -939,9 +942,9 @@ class TestListPrsAuthorScope:
             ]
             return MagicMock(stdout=json.dumps(payload))
 
-        monkeypatch.setattr(fleet_sync_module, "_gh", fake_gh)
+        monkeypatch.setattr(fleet_pr_api, "_gh", fake_gh)
 
-        prs = fleet_sync_module.list_prs("RepoA", "HomericIntelligence")
+        prs = fleet_pr_api.list_prs("RepoA", "HomericIntelligence")
 
         assert [p.number for p in prs] == [1]
 
@@ -954,24 +957,24 @@ class TestAsciiFlag:
         from dataclasses import FrozenInstanceError
 
         with pytest.raises(FrozenInstanceError):
-            fleet_sync_module.ASCII_SYMBOLS.check = "X"
+            cast(Any, fleet_models.ASCII_SYMBOLS).check = "X"
 
     def test_presets_have_expected_glyphs(self) -> None:
         """Unicode and ASCII symbol presets have the correct glyphs."""
-        assert fleet_sync_module.UNICODE_SYMBOLS.check == "✓"
-        assert fleet_sync_module.UNICODE_SYMBOLS.banner == "══"
-        assert fleet_sync_module.UNICODE_SYMBOLS.arrow == "→"
-        assert fleet_sync_module.UNICODE_SYMBOLS.dash == "—"
-        assert fleet_sync_module.ASCII_SYMBOLS.check == "*"
-        assert fleet_sync_module.ASCII_SYMBOLS.banner == "=="
-        assert fleet_sync_module.ASCII_SYMBOLS.arrow == "->"
-        assert fleet_sync_module.ASCII_SYMBOLS.dash == "--"
+        assert fleet_models.UNICODE_SYMBOLS.check == "✓"
+        assert fleet_models.UNICODE_SYMBOLS.banner == "══"
+        assert fleet_models.UNICODE_SYMBOLS.arrow == "→"
+        assert fleet_models.UNICODE_SYMBOLS.dash == "—"
+        assert fleet_models.ASCII_SYMBOLS.check == "*"
+        assert fleet_models.ASCII_SYMBOLS.banner == "=="
+        assert fleet_models.ASCII_SYMBOLS.arrow == "->"
+        assert fleet_models.ASCII_SYMBOLS.dash == "--"
 
     def test_process_repo_emits_ascii_banner_when_ascii_symbols_passed(
         self, monkeypatch: pytest.MonkeyPatch, tmp_path: Path, capture_fleet_sync_logs
     ) -> None:
         """process_repo logs ASCII banner under ASCII_SYMBOLS — no module state."""
-        monkeypatch.setattr(fleet_sync_module, "list_prs", lambda _repo, _org: [])
+        monkeypatch.setattr(fleet_coordinator, "list_prs", lambda _repo, _org: [])
 
         import argparse
 
@@ -985,12 +988,12 @@ class TestAsciiFlag:
                 ascii=True,
             )
 
-            fleet_sync_module.process_repo(
+            fleet_coordinator.process_repo(
                 "test-repo",
                 "HomericIntelligence",
                 args,
                 tmp_path,
-                symbols=fleet_sync_module.ASCII_SYMBOLS,
+                symbols=fleet_models.ASCII_SYMBOLS,
             )
 
             # Check logged messages
@@ -1002,7 +1005,7 @@ class TestAsciiFlag:
         self, monkeypatch: pytest.MonkeyPatch, tmp_path: Path, capture_fleet_sync_logs
     ) -> None:
         """Default kwarg gives Unicode — backward-compat for existing callers."""
-        monkeypatch.setattr(fleet_sync_module, "list_prs", lambda _repo, _org: [])
+        monkeypatch.setattr(fleet_coordinator, "list_prs", lambda _repo, _org: [])
 
         import argparse
 
@@ -1016,7 +1019,7 @@ class TestAsciiFlag:
                 ascii=False,
             )
 
-            fleet_sync_module.process_repo("test-repo", "HomericIntelligence", args, tmp_path)
+            fleet_coordinator.process_repo("test-repo", "HomericIntelligence", args, tmp_path)
 
             # Check logged messages
             output = "\n".join(logged_messages)
@@ -1029,7 +1032,7 @@ class TestAsciiFlag:
         failing — the behavioral tests pass ``symbols=`` directly and bypass
         argparse entirely.
         """
-        parser = fleet_sync_module._build_parser()
+        parser = fleet_cli._build_parser()
 
         ascii_action = next((a for a in parser._actions if "--ascii" in a.option_strings), None)
         assert ascii_action is not None, "--ascii flag is not registered on the parser"
@@ -1046,7 +1049,7 @@ class TestAsciiFlag:
         Regression guard for the self-contradictory ``== for ==`` mapping that
         conveyed nothing about what the flag swaps.
         """
-        parser = fleet_sync_module._build_parser()
+        parser = fleet_cli._build_parser()
         ascii_action = next(a for a in parser._actions if "--ascii" in a.option_strings)
         help_text = ascii_action.help or ""
